+++ tests/FeatureTests/Endpoint/System/GetWellKnownSecurityTxtTest.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\tests\FeatureTests\Endpoint\System;
+
+use App\Tests\FeatureTests\BaseRequestTestCase;
+
+class GetWellKnownSecurityTxtTest extends BaseRequestTestCase
+{
+    public function testGetWellKnownSecurityTxt(): void
+    {
+        $response = $this->runGetRequest('/.well-known/security.txt', null);
+        $this->assertIsTextResponse($response, 200);
+    }
+}